Sync the display to real time instead of the audio player
Instead of asking the audio player what time its playing, we
remember the time of day when we start playing and synch to that,
which makes for amoother scrolling. If our idea of what time is
playing get more than a 20th of a secnod out of synch with what the
audio player thinks, we resynchronize our real-time idea to what it says.
With SDL, in practice, on this system, this happens abuot once a minute.
W

+++ audio_cache.c
@@ -65,7 +65,7 @@ static off_t audio_cache_size = 0;	/* Size of cache in sample frames */
 int
 read_cached_audio(char *data,
 		  af_format_t format, int channels,
-		  int start, int frames_to_read)
+		  off_t start, int frames_to_read)
 {
     int frames_written = 0;
     size_t framesize;
